requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-allauth==0.54.0
requests==2.31.0